Conflicts:
	Config.ini
	src/Personas/Clientes/ControladorClientes.java
	src/Personas/Clientes/frmMostrarClientes.java

+++ src/Personas/Clientes/ControladorClientes.java
@@ -28,11 +28,16 @@ public class ControladorClientes {
         }
     }
 
-    public void eliminar(frmRegistrarCliente vista) throws ErrorConexion {
+    public void eliminar(frmBuscarPersona vista) throws ErrorConexion {
         cliente = new clsClientes();
-        cliente.setCedula(vista.getCedula());
-        BD = new BaseDatos("DELETE FROM tblclientes WHERE Cedula = ?");
-        BD.ejecutar(new Object[]{cliente.getCedula()});
+        cliente.setCedula(vista.getPersonas().getValueAt(vista.getPersonas().getSelectedRow(), 0).toString());
+        if (ErrorMensaje.mostrarMensajes()) {
+            JOptionPane.showMessageDialog(vista, ErrorMensaje.getMsj(), "Error", 0);
+        } else {
+            BD = new BaseDatos("DELETE FROM tblclientes WHERE Cedula = ?");
+            BD.ejecutar(new Object[]{cliente.getCedula()});
+            JOptionPane.showMessageDialog(vista, "El Registro se Elimino", "Eliminar Cliente", 1);
+        }
     }
 
     public void filtar(frmBuscarPersona vista) throws ErrorConexion {
@@ -66,18 +71,20 @@ public class ControladorClientes {
     }
 
     public void cargarClientes(frmMostrarClientes vista) throws ErrorConexion {
-        BD = new BaseDatos("SELECT cliente.Cedula, cliente.Nombre, cliente.`Fecha Nac`, cliente.Telefono, cliente.Correo FROM tblclientes AS cliente INNER JOIN tblpruebas as prueba on prueba.IdCliente = cliente.Cedula");
+        BD = new BaseDatos("SELECT cliente.Cedula, cliente.Nombre, cliente.`Fecha Nac`, cliente.Telefono, "
+                + "cliente.Correo FROM tblclientes AS cliente INNER JOIN tblpruebas as prueba on cliente.Cedula = prueba.IdCliente");
         BD.ejecutar();
         DefaultTableModel modelo = (DefaultTableModel) vista.getTblClientes().getModel();
         modelo.setNumRows(0);
         Object obj[];
+        int x = 0;
         do {
             obj = BD.getObjet();
             if (obj != null) {
-                 cliente = new clsClientes(obj);
+                cliente = new clsClientes(obj);
                 if (validar(vista, cliente)) {
-                      modelo.addRow(cliente.toObjects());
-                } 
+                    modelo.addRow(cliente.toObjects());
+                }
             }
         } while (obj != null);
     }
@@ -97,8 +104,8 @@ public class ControladorClientes {
         cliente = new clsClientes(obj);
         return cliente;
     }
-    
-    private boolean validar(frmMostrarClientes vista, clsClientes cliente) {
+
+    public boolean validar(frmMostrarClientes vista, clsClientes cliente) {
         int fila = vista.getTblClientes().getRowCount();
         for (int i = 0; i < fila; i++) {
             if (vista.getTblClientes().getValueAt(i, 0).toString().equals(cliente.getCedula())) {
@@ -107,4 +114,11 @@ public class ControladorClientes {
         }
         return true;
     }
+
+    public int comprobarPrueba(frmMostrarClientes vista) throws ErrorConexion {
+        BD = new BaseDatos("SELECT count(Id) FROM tblpruebas WHERE IdCliente = ?");
+        BD.ejecutar(new Object[]{vista.getTblClientes().getValueAt(vista.getTblClientes().getSelectedRow(), 0).toString()});
+        return (int) BD.getObjet()[0];
+    }
+
 }
